createCart and deleteCart completed

+++ src/main/java/com/mr/onlineshopping/services/CartService.java
@@ -2,12 +2,19 @@ package com.mr.onlineshopping.services;
 
 import com.mr.onlineshopping.entity.Article;
 import com.mr.onlineshopping.entity.Cart;
+import com.mr.onlineshopping.entity.User;
 import com.mr.onlineshopping.exceptions.CartNotFound;
+import com.mr.onlineshopping.exceptions.UserNotFound;
 import com.mr.onlineshopping.interfaces.CartFunctions;
 import com.mr.onlineshopping.repository.CartRepository;
+import com.mr.onlineshopping.repository.UserRepository;
+import jakarta.persistence.EntityManager;
+import jakarta.persistence.PersistenceContext;
+import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.Optional;
 import java.util.Set;
@@ -17,6 +24,9 @@ public class CartService implements CartFunctions {
 
     @Autowired
     CartRepository cartRepository;
+
+    @Autowired
+    UserRepository userRepository;
 
     @Override
     public Optional<Cart> getCartById(int cartId) {
@@ -55,12 +65,24 @@ public class CartService implements CartFunctions {
     }
 
     @Override
+    public boolean createCart(int userId) throws UserNotFound {
+        User user = userRepository.findById(userId).orElseThrow(() -> new UserNotFound(userId));
+        Cart cart = new Cart();
+        cart.setUser(user);
+        cart.setTotalPrice(BigDecimal.valueOf(0)); // Dovrebbe essere di default 0.00, ma mi obbliga a settarlo
+        cartRepository.save(cart);
+        return true;
+    }
+
+    @Override
     public boolean clearCart(int cartId) {
         return false;
     }
 
+    @Transactional
     @Override
-    public boolean deleteCart(int cartId) {
+    public boolean deleteCart(int cartId) throws CartNotFound {
+        if (!cartRepository.existsById(cartId)) { throw new CartNotFound(cartId); }
         cartRepository.deleteById(cartId);
         return true;
     }
